Refactor code to improve data handling and file import
Weight and price in front_interaction.py models have been changed from float to Decimal for more accurate data handling. Rename func.py to components/func.py, simplifying codes by adding a new function for writing DataFrame to SQL. Also, restructuring file importing routes and abstracting file handling tasks into components/datafiles.py, providing better scalability and easier maintenance.

+++ models/front_interaction.py
@@ -41,8 +41,8 @@ class Invoice(MyModel, BaseModel):
     client: str
     route: str
     cargo: str
-    weight: float
-    price: float
+    weight: Decimal
+    price: Decimal
     departure_date: datetime
     arrival_date: datetime
